+++ funding/get_internal_funded_tx.js
@@ -1,16 +1,13 @@
 const asyncAuto = require('async/auto');
 const {fundPsbt} = require('ln-service');
-const {getChainFeeRate} = require('ln-service');
 const {returnResult} = require('asyncjs-util');
 const {signPsbt} = require('ln-service');
 const {Transaction} = require('bitcoinjs-lib');
 
-const fastConfirmationTarget = 2;
-const {floor} = Math;
+const askForFeeRate = require('./ask_for_fee_rate');
+
 const {fromHex} = Transaction;
 const {isArray} = Array;
-const maxMultiplier = 100;
-const slowConfirmationTarget = 1e3;
 
 /** Get a funded tx using internal LND funds
 
@@ -56,82 +53,22 @@ module.exports = (args, cbk) => {
         return cbk();
       },
 
-      // Get the default chain fee rate
-      getDefaultChainFeeRate: ['validate', ({}, cbk) => {
-        return getChainFeeRate({lnd: args.lnd}, cbk);
-      }],
-
-      // Get the maximum chain fee rate
-      getMaxChainFeeRate: ['validate', ({}, cbk) => {
-        return getChainFeeRate({
-          confirmation_target: fastConfirmationTarget,
-          lnd: args.lnd,
-        },
-        cbk);
-      }],
-
-      // Get the minimum chain fee rate
-      getMinChainFeeRate: ['validate', ({}, cbk) => {
-        return getChainFeeRate({
-          confirmation_target: slowConfirmationTarget,
-          lnd: args.lnd,
-        },
-        cbk);
-      }],
-
       // Ask for fee rate to use for internal funding
-      askForFeeRate: [
-        'getDefaultChainFeeRate',
-        'getMaxChainFeeRate',
-        'getMinChainFeeRate',
-        ({
-          getDefaultChainFeeRate,
-          getMaxChainFeeRate,
-          getMinChainFeeRate,
-        },
-        cbk) =>
-      {
+      askForFeeRate: ['validate', ({}, cbk) => {
         // Exit early when the fee rate was already set
         if (!!args.chain_fee_tokens_per_vbyte) {
-          return cbk(null, args.chain_fee_tokens_per_vbyte);
+          return cbk(null, {
+            tokens_per_vbyte: args.chain_fee_tokens_per_vbyte,
+          });
         }
 
-        return args.ask({
-          default: getDefaultChainFeeRate.tokens_per_vbyte,
-          message: 'Chain fee per vbyte?',
-          name: 'rate',
-          type: 'number',
-        },
-        ({rate}) => cbk(null, rate));
+        return askForFeeRate({ask: args.ask, lnd: args.lnd}, cbk);
       }],
 
       // Create a PSBT to the outputs
-      createPsbt: [
-        'askForFeeRate',
-        'getMaxChainFeeRate',
-        'getMinChainFeeRate',
-        ({
-          askForFeeRate,
-          getMaxChainFeeRate,
-          getMinChainFeeRate,
-        },
-        cbk) =>
-      {
-        const hasFee = !!args.chain_fee_tokens_per_vbyte;
-
-        const maxFee = getMaxChainFeeRate.tokens_per_vbyte * maxMultiplier;
-        const minFee = floor(getMinChainFeeRate.tokens_per_vbyte);
-
-        if (!hasFee && askForFeeRate > maxFee) {
-          return cbk([400, 'MaxFeePerVbyteExceeded', {maximum: maxFee}]);
-        }
-
-        if (!hasFee && (!askForFeeRate || askForFeeRate < minFee)) {
-          return cbk([400, 'ExpectedHigherMinFeePerVbyte', {minimum: minFee}]);
-        }
-
+      createPsbt: ['askForFeeRate', ({askForFeeRate}, cbk) => {
         return fundPsbt({
-          fee_tokens_per_vbyte: askForFeeRate,
+          fee_tokens_per_vbyte: askForFeeRate.tokens_per_vbyte,
           lnd: args.lnd,
           outputs: args.outputs,
         },