+++ funding/index.js
@@ -1,4 +1,5 @@
+const askForFeeRate = require('./ask_for_fee_rate');
 const getFundedTransaction = require('./get_funded_transaction');
 const getPsbtFromTransaction = require('./get_psbt_from_transaction');
 
-module.exports = {getFundedTransaction, getPsbtFromTransaction};
+module.exports = {askForFeeRate, getFundedTransaction, getPsbtFromTransaction};